+++ src/rifgen/mod.rs
@@ -1,6 +1,7 @@
 pub mod context;
 pub mod description;
 pub mod interrupt;
+pub mod logic_expr;
 pub mod field;
 pub mod register;
 pub mod page;
@@ -9,4 +10,15 @@ pub mod rif;
 pub mod rifmux;
 pub mod order_dict;
 
-pub use {context::*, description::*, interrupt::*, field::*, register::*, page::*, rif::*, rifmux::*, width::*};
+pub use {
+	context::*,
+	description::*,
+	interrupt::*,
+	field::*,
+	register::*,
+	logic_expr::*,
+	page::*,
+	rif::*,
+	rifmux::*,
+	width::*
+};